+++ examples/using-typescript/gatsby-config.ts
@@ -1,0 +1,11 @@
+import type { GatsbyConfig } from "gatsby"
+
+const config: GatsbyConfig = {
+  siteMetadata: {
+    siteName: `Using TypeScript`,
+    sourceUrl: `https://github.com/gatsbyjs/gatsby/tree/master/examples/using-typescript`,
+  },
+  plugins: [],
+}
+
+export default config