+++ examples/using-vanilla-extract/gatsby-config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  siteMetadata: {
-    title: `using-vanilla-extract`,
-    siteUrl: `https://www.yourdomain.tld`,
-  },
-  plugins: [`gatsby-plugin-vanilla-extract`]
-}