+++ packages/gatsby-source-wordpress/src/steps/index.ts
@@ -19,3 +19,4 @@ export {
 
 export { pluginOptionsSchema } from "~/steps/declare-plugin-options-schema"
 export { logPostBuildWarnings } from "~/steps/log-post-build-warnings"
+export { imageRoutes } from "~/steps/image-routes"