+++ packages/gatsby-source-wordpress/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.9.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-wordpress@6.9.0/packages/gatsby-source-wordpress) (2022-03-01)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.9)
+
+#### Chores
+
+- Update Contributing section, RFC process, Governance Model [#34900](https://github.com/gatsbyjs/gatsby/issues/34900) ([4cb514f](https://github.com/gatsbyjs/gatsby/commit/4cb514f7a545a0afa0b718e07f7c75688122f225))
+- Format changelog files ([088f23b](https://github.com/gatsbyjs/gatsby/commit/088f23b084b67f746a383e06e9216cef83270317))
+
+### [6.8.2](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-wordpress@6.8.2/packages/gatsby-source-wordpress) (2022-03-01)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
+### [6.8.1](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-wordpress@6.8.1/packages/gatsby-source-wordpress) (2022-02-25)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
 ## [6.8.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-wordpress@6.8.0/packages/gatsby-source-wordpress) (2022-02-22)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.8)